+++ src/Fable.Core/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -769,7 +770,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.7.5-alpha.1 - 2016-10-16
 
-* Fix prerelease semver. See: https://docs.npmjs.com/misc/semver#prerelease-tags
+* Fix prerelease semver. [See documentation](https://docs.npmjs.com/misc/semver#prerelease-tags)
 
 ## 0.7.5-alpha - 2016-10-16