+++ src/Fable.Core/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* [JS] Add `ReadonlyArray` to `Fable.Core.JS` by (@MangelMaxime)
+
 ## 4.3.0 - 2024-01-25
 
 ### Added