+++ src/Fable.Core/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 4.4.0 - 2025-02-16
+
 ### Added
 
 * [TS/JS] Add `Pojo` attribute (by @alfonsogarciacaro)